requirements.txt
python-dotenv
groq
pandas
openpyxl
fastapi
uvicorn
python-multipart
RestrictedPython